+++ front/pages/api/w/[wId]/vaults/[vId]/members.ts
@@ -4,7 +4,6 @@ import { isLeft } from "fp-ts/lib/Either";
 import * as reporter from "io-ts-reporters";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import { updateVaultPermissions } from "@app/lib/api/vaults";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/wrappers";
 import type { Authenticator } from "@app/lib/auth";
 import { DustError } from "@app/lib/error";
@@ -71,9 +70,8 @@ async function handler(
         });
       }
 
-      const updateRes = await updateVaultPermissions(
+      const updateRes = await vault.updatePermissions(
         auth,
-        vault,
         bodyValidation.right
       );
       if (updateRes.isErr()) {
